Create the ControllerManager after robot was openned
In the khi_robot_control node, the ControllerManager (cm) provides
services that require controller classes to be loaded beforehand.
This loading occurs when we open the robot.

By moving the creation of the ControllerManager after we try to open
the robot, we ensure that no race condition will occur and that the
controllers will always be loaded when the provided services are called.